refactor(main/profile): keep info card decomposition local
